Remove an invalid assertion
This assertion is invalid.

If you have a HeightsLeaf with positions == [5, 10, 15], by 0..<5, you get [5, 5], not [5]. This felt a bit wrong to me, but I ran into issues trying to fix it, and we take care of appending to leafs that end with an empty blank line in HeightsLeaf.pushMaybeSplitting(other:), so this doesn’t seem like it’s going to be a problem (famous last words).

Fixes #55