Refine behavior when expanding to a new granularity
When deciding which granule to select when double clicking, we want to use  the glyph’s caret offset to determine the index, rather than the glyph’s halfway point.

In other words, in "   foo", double clicking on left edge of the “f” should select “foo”, but double clicking one point to the left should select “   “.

The purpose is to make it easier to select the space in a string like “foo bar”. You can double click anywhere in the space to select it.