Fix IntervalCache invalidation at line boundaries
If there are two contiguous lines in the IntervalCache (e.g. 10..<20 and 20..<30), and have a delta that removes up through the end interval (e.g. removeSubrange(18..<20), we should invalidate 10..<20 and 20..<30.

We need to do this because index 19 could have been a “\n”, which would merge the two lines, and thus invalidate 20..<30.